Fix Subject Sharing
According to the spec, replies should have the same subject. For simplicity, when we create a child object, we make an extra call (to the parent) to get the subject, and use that for the subject of the child message.

This PR also includes fixes for providing the parent id in other areas